Check for type

+++ proxy/proxy.py
@@ -206,7 +206,8 @@ class Proxy:
         self._delete_pid_file()
 
     def _write_pid_file(self) -> None:
-        if self.flags.pid_file is not None:
+        if self.flags.pid_file is not None \
+                and type(self.flags.pid_file) == str:
             if Path(self.flags.pid_file).exists():
                 raise ValueError(
                     'pid file {0} already exists'.format(
@@ -217,7 +218,9 @@ class Proxy:
                 pid_file.write(bytes_(os.getpid()))
 
     def _delete_pid_file(self) -> None:
-        if self.flags.pid_file and os.path.exists(self.flags.pid_file):
+        if self.flags.pid_file \
+            and type(self.flags.pid_file) == str \
+                and os.path.exists(self.flags.pid_file):
             os.remove(self.flags.pid_file)
 
     @property
